GH-90 Moved dialogs to specific namespace.
